TW-249: Find deadcode. Refactor
