+++ src/app/atoms/index.ts
@@ -18,3 +18,5 @@ export { NoSpaceField } from './NoSpaceField';
 export { TabSwitcher } from './TabSwitcher';
 
 export { SeedPhraseInput } from './SeedPhraseInput';
+
+export { Stepper } from './Stepper/Stepper';
