+++ src/app/pages/NewWallet/import/ImportFromKeystoreFile.tsx
@@ -8,7 +8,7 @@ import { ReactComponent as TrashbinIcon } from 'app/icons/bin.svg';
 import { ReactComponent as PaperclipIcon } from 'app/icons/paperclip.svg';
 import { T, t } from 'lib/i18n/react';
 import { decryptKukaiSeedPhrase } from 'lib/temple/front';
-import { AlertFn, useAlert } from 'lib/ui/dialog';
+import { AlertFn, useAlert } from 'lib/ui';
 
 interface FormData {
   keystoreFile?: FileList;
