+++ src/lib/ui/index.ts
@@ -1,0 +1,2 @@
+export type { AlertFn } from './dialog';
+export { useAlert } from './dialog';
